# build/build/swagger.json
{
    "components": {
...
        "schemas": {
...
            "TaskTypeEnum": {
                "description": "Тип задачи",
                "enum": [
                    "task",
                    "bug"
                ],
                "type": "string"
            },
            "TaskStatusEnum": {
                "description": "Статус",
                "enum": [
                    "opened",
                    "inProgress",
                    "testing",
                    "complete"
                ],
                "type": "string"
            },
            "TaskRankEnum": {
                "description": "Приоритет",
                "enum": [
                    "low",
                    "medium",
                    "high"
                ],
                "type": "string"
            },
            "TaskDTO": {
                "description": "Модель задачи",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "id задачи"
                    },
                    "userId": {
                        "type": "string",
                        "description": "id пользователя"
                    },
                    "assignedId": {
                        "type": "string",
                        "description": "id пользователя на которого назначена задача"
                    },
                    "title": {
                        "type": "string",
                        "description": "Заголовок"
                    },
                    "description": {
                        "type": "string",
                        "description": "Описание"
                    },
                    "type": {
                        "$ref": "#/components/schemas/TaskTypeEnum",
                        "description": "Тип задачи"
                    },
                    "dateOfCreation": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Дата создания"
                    },
                    "dateOfUpdate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Дата обновления"
                    },
                    "timeInMinutes": {
                        "type": "number",
                        "format": "double",
                        "description": "Затраченное время"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TaskStatusEnum",
                        "description": "Статус"
                    },
                    "rank": {
                        "$ref": "#/components/schemas/TaskRankEnum",
                        "description": "Приоритет"
                    }
                },
                "required": [
                    "userId",
                    "assignedId",
                    "title",
                    "description",
                    "type",
                    "rank"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ListDTO_TaskDTO_": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/TaskDTO"
                        },
                        "type": "array",
                        "description": "Результат"
                    },
                    "page": {
                        "type": "number",
                        "format": "double",
                        "description": "Номер страницы"
                    },
                    "limit": {
                        "type": "number",
                        "format": "double",
                        "description": "Кол-во элементов на странице"
                    },
                    "total": {
                        "type": "number",
                        "format": "double",
                        "description": "Всего записей по фильтру"
                    }
                },
                "required": [
                    "data",
                    "page",
                    "limit",
                    "total"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "TaskFilter": {
                "properties": {
                    "query": {
                        "type": "string",
                        "description": "Заголовок"
                    },
                    "assignedUsers": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "description": "id исполнителя"
                    },
                    "userIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "description": "id создателя задачи"
                    },
                    "type": {
                        "items": {
                            "$ref": "#/components/schemas/TaskTypeEnum"
                        },
                        "type": "array",
                        "description": "Тип задачи"
                    },
                    "status": {
                        "items": {
                            "$ref": "#/components/schemas/TaskStatusEnum"
                        },
                        "type": "array",
                        "description": "Статус задачи"
                    },
                    "rank": {
                        "items": {
                            "$ref": "#/components/schemas/TaskRankEnum"
                        },
                        "type": "array",
                        "description": "Приоритет задачи"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "QueryDTO_TaskFilter_": {
                "properties": {
                    "filter": {
                        "$ref": "#/components/schemas/TaskFilter",
                        "description": "Фильтр"
                    },
                    "page": {
                        "type": "number",
                        "format": "double",
                        "description": "Номер страницы (начиная с нуля)"
                    },
                    "limit": {
                        "type": "number",
                        "format": "double",
                        "description": "Кол-во элементов на странице"
                    }
                },
                "required": [
                    "filter",
                    "page",
                    "limit"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "UserDTO": {
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "id пользователя"
                    },
                    "login": {
                        "type": "string",
                        "description": "Логин"
                    },
                    "username": {
                        "type": "string",
                        "description": "Имя"
                    },
                    "about": {
                        "type": "string",
                        "nullable": true,
                        "description": "О себе"
                    },
                    "photoUrl": {
                        "type": "string",
                        "nullable": true,
                        "description": "Ссылка на фотографию"
                    }
                },
                "required": [
                    "login",
                    "username"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ListDTO_UserDTO_": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/UserDTO"
                        },
                        "type": "array",
                        "description": "Результат"
                    },
                    "page": {
                        "type": "number",
                        "format": "double",
                        "description": "Номер страницы"
                    },
                    "limit": {
                        "type": "number",
                        "format": "double",
                        "description": "Кол-во элементов на странице"
                    },
                    "total": {
                        "type": "number",
                        "format": "double",
                        "description": "Всего записей по фильтру"
                    }
                },
                "required": [
                    "data",
                    "page",
                    "limit",
                    "total"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UserFilter": {
                "properties": {
                    "query": {
                        "type": "string",
                        "nullable": true,
                        "description": "Имя пользоваетля"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "QueryDTO_UserFilter_": {
                "properties": {
                    "filter": {
                        "$ref": "#/components/schemas/UserFilter",
                        "description": "Фильтр"
                    },
                    "page": {
                        "type": "number",
                        "format": "double",
                        "description": "Номер страницы (начиная с нуля)"
                    },
                    "limit": {
                        "type": "number",
                        "format": "double",
                        "description": "Кол-во элементов на странице"
                    }
                },
                "required": [
                    "filter",
                    "page",
                    "limit"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "LoginDTO": {
                "description": "Модель авторизации пользователя",
                "properties": {
                    "login": {
                        "type": "string",
                        "description": "Логин"
                    },
                    "password": {
                        "type": "string",
                        "description": "Пароль"
                    }
                },
                "required": [
                    "login",
                    "password"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "school-final-task-service",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/tasks": {
            "post": {
                "operationId": "GetByFilter",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListDTO_TaskDTO_"
                                }
                            }
                        }
                    }
                },
                "description": "Получение списка задач по фильтру",
                "tags": [
                    "Tasks"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueryDTO_TaskFilter_"
                            }
                        }
                    }
                }
            }
        },
...
        "/users": {
            "post": {
                "operationId": "GetByFilter",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListDTO_UserDTO_"
                                }
                            }
                        }
                    }
                },
                "description": "Получение списка пользователей по фильтру",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueryDTO_UserFilter_"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "operationId": "GetById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Текст ошибки"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Получение пользователя по id",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/users/edit": {
            "put": {
                "operationId": "Edit",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Текст ошибки"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Текст ошибки"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Редактирование пользователя",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserDTO"
                                    },
                                    {
                                        "$ref": "#/components/schemas/LoginDTO"
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "operationId": "Login",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Текст ошибки"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Auth",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Текст ошибки"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Авторизация пользователя",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDTO"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "delete": {
                "operationId": "Delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                },
                                "examples": {
                                    "Example 1": {
                                        "value": {
                                            "message": "Текст ошибки"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Удаление пользователя по id",
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "description": "id пользователя",
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        }
    },
    "servers": [
        {
            "url": "/api"
        }
    ]
}